edited the structure to solve possiable conflicts
