--- package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "express-validator": "^6.14.0"
-  }
-}
